Create [hard] 587. Erect the Fence (Convex Hull)
problem description: You are given an array trees where trees[i] = [xi, yi] represents the location of a tree in the garden.

You are asked to fence the entire garden using the minimum length of rope as it is expensive. The garden is well fenced only if all the trees are enclosed.

Return the coordinates of trees that are exactly located on the fence perimeter.


Explanation: So, this algorithm works by three main steps: First, we find the minimum point M, which is guaranteed to be on the hull, since it is one of the outermost points; the hull is the set of outermost points which, when connected by straight lines to their nearest neighbors in the hull, entraps the most points from the given points. Second, we sort the points according to the angle they form with M, hence we use the tangent function atan2(y, x), which is the inverse tangent function (if you recall from trigonometry, the inverse tangent tells us the angle formed between two lines, one line (y) being on a side opposite to the angle and the other line being adjacent (x) to the angle). This ensures that we loop through the points in an order that prioritizes points on the hull. Thirdly, akin to a monotonic stack, we loop through the points collecting points that form a line L (stack[-2], stack[-1]) with the leading point on the stack such that L is counter-clockwise with respect to the line L' of the two previous points (stack[-3], stack[-2]) on the stack; if L' has to turn counterclockwise to overlap L, then we keep L, if not, we pop the leading point (stack[-2]) which makes L' from the stack and continue popping until L' has to turn counterclockwise to lay on top of L. We determine this clockwise/counter-clockwise rotation of L and L' by taking the determinant of the matrix formed by the vectors which define L and L'; if the area is negative, then L' turns clockwise. Counter-clockwise turns imply that points are trapped inside of the hull.