Update [hard] 827. Making A Large Island
Description: You are given an n x n binary matrix grid. You are allowed to change at most one 0 to be 1.

Return the size of the largest island in grid after applying this operation.

An island is a 4-directionally connected group of 1s.

Explanation: So, we have multiples sets of disjoint islands and we are allowed to paint one point in the matrix such that multiples islands can become connected. So, first we just loop over the entire matrix looking for islands. If we see an island, then we enter a depth first search (dfs) which paints the entire island a particular color (a unique integer) which we use as a key in the dictionary (ie, colors) for storing the size of the islands. Then, on another loop through the whole matrix, we now look for zeros. At each zero, we see if there are islands to the left, right, up, or down. If so, grab the key which identifies the island; then, since all of these islands (identified by the keys we just collected) are disjoint by a single zero in the matrix, we can connect them by coloring the zero--then, the size of the island is the (sum of the sizes of the island) + 1, +1 because of the zero that we just painted (ie, turned into land). Then we just return the size of the largest island we found. Also, we used a set to collect the keys because it is possible that we grab the same key multiple times (eg, imagine an island shaped like L and we are in the white space [a zero] just above the bottom left corner of the L).