Rename 85. Maximal Rectangle to [hard] 85. Maximal Rectangle
Description: Given a rows x cols binary matrix filled with 0's and 1's, find the largest rectangle containing only 1's and return its area.

Explanation: So, an insight you can have that will make this problem easy is that we can do prefix sum with the rows or columns (I did columns) to store the heights of the columns in the matrix. This allows us to run through the rows in the next loop and then take the best rectangle size.