Create [hard, IOI 98] 711. Number of Distinct Islands II
description: You are given an m x n binary matrix grid. An island is a group of 1's (representing land) connected 4-directionally (horizontal or vertical.) You may assume all four edges of the grid are surrounded by water.

An island is considered to be the same as another if they have the same shape, or have the same shape after rotation (90, 180, or 270 degrees only) or reflection (left/right direction or up/down direction).

Return the number of distinct islands. 

http://olympiads.win.tue.nl/ioi/ioi98/contest/day1/starry/starry.html

Explanation: So, this is a harder problem; it was even included in the 1998 IOI. The first thing we do is go through the islands and then call the dfs function whenever we see an island. The dfs function I wrote is quite long. The objective of this function is to store each island in all possible forms: all possible rotations and reflections of the island. We throw these rotations and reflections into tuples, sort those tuples, and then put them in a set--we have to make them tuples so that they can be hashed (sets do not accept lists). Further, I sort these tuples so that if we encounter the same island later, we do not double count it (eg, if we first encounter an island shaped like 6 and then an island shaped like 9, then we want to only count these islands once, but the order of their rotations will be different--so, to ignore the difference in ordering of rotations, I simply sort the list of rotations). Then I just return the length of the set of islands.