Create [hard] 1301. Number of Paths with Max Score
Description: You are given a square board of characters. You can move on the board starting at the bottom right square marked with the character 'S'.

You need to reach the top left square marked with the character 'E'. The rest of the squares are labeled either with a numeric character 1, 2, ..., 9 or with an obstacle 'X'. In one move you can go up, left or up-left (diagonally) only if there is no obstacle there.

Return a list of two integers: the first integer is the maximum sum of numeric characters you can collect, and the second is the number of such paths that you can take to get that maximum sum, taken modulo 10^9 + 7.

In case there is no path, return [0, 0].

Explanation: So, we can solve both objectives of this problem in parallel. We need to find the number of paths through the matrix that yield the maximum score and we need to find the maximum score. So, to find the number of paths in a matrix wherein we can only move down, right and down-right, we can actually just count the number of paths to get to each coordinate path(i, j) and then path(i+1, j+1) = path(i, j) + path(i+1, j) + path(i, j+1) since those three coordinates are the only cells from which we can enter (i+1, j+1). But we want to find the number of paths that have the maximum sum. Therefore, we only add paths to the current cell's number_of_ways if the score in the previous reachable cell has the highest score from the three previous cells (so, we'd add all three if they have equal scores). Then we just return the max score and the number of ways to get that max score.