Update [hard] 124. Binary Tree Maximum Path Sum
description: A path in a binary tree is a sequence of nodes where each pair of adjacent nodes in the sequence has an edge connecting them. A node can only appear in the sequence at most once. Note that the path does not need to pass through the root.

The path sum of a path is the sum of the node's values in the path.

Given the root of a binary tree, return the maximum path sum of any path.

Explanation: https://leetcode.com/problems/binary-tree-maximum-path-sum/discuss/1414070/Python-10-simple-lines-explained-with-links-to-2-similar-problems