Create [hard] 968. Binary Tree Cameras
Description: You are given the root of a binary tree. We install cameras on the tree nodes where each camera at a node can monitor its parent, itself, and its immediate children.

Return the minimum number of cameras needed to monitor all nodes of the tree.

Explanation: So, we go through the bst in the order of left, right, middle. At each node, we ask if either of its kids are NOT painted. If not, then we paint the node and the kids and parent node.