Create [hard] 834. Sum of Distances in Tree
Description: There is an undirected connected tree with n nodes labeled from 0 to n - 1 and n - 1 edges.

You are given the integer n and the array edges where edges[i] = [ai, bi] indicates that there is an edge between nodes ai and bi in the tree.

Return an array answer of length n where answer[i] is the sum of the distances between the ith node in the tree and all other nodes.

Explanation: so, the first thing we do is a dfs to find the number of edges between the root (which we assume is 0, since we can use any node as the root, so may as well just use 0), and we get the size of every sub-tree of every node, which is used in the second step. Then, in the second dfs, which is done with a stack, we take the answer from the root and modify this answer such that it becomes the answer for each child. This is done by observing that, when we move from one node to its child, every node in the sub-tree of the child gets 1 closer and every node in the sub-tree of the parent, disregarding the child we are getting the answer for, gets 1 farther. Hence, we have line 31.