Create [hard] 1293. Shortest Path in a Grid with Obstacles Elimination
You are given an m x n integer matrix grid where each cell is either 0 (empty) or 1 (obstacle). You can move up, down, left, or right from and to an empty cell in one step.

Return the minimum number of steps to walk from the upper left corner (0, 0) to the lower right corner (m - 1, n - 1) given that you can eliminate at most k obstacles. If it is not possible to find such walk return -1.

Explanation: So, this is a BFS problem. We are trying to find the shortest path to the end of the matrix which involves breaking k or less walls. To do this, we just take every possible trajectory, wherein each state in the trajectory is determined by 3 parameters, namely those we store in the visited set (row position, column position, and the number of walls that have been broken so far). When we find a trajectory that makes it to the end of the matrix, then that must be the shortest path, since it is the first trajectory which made it to the end. And we deal with the wall-breaking constraint by throwing away any trajectories that have broken too many walls.